+++ from_json/as.ts
@@ -3,10 +3,10 @@
 import { FromJSONStrategy, JSONValue, Serializable } from "../serializable.ts";
 
 /** revive data using `fromJSON` on a subclass type */
-export function fromJSONAs<T>(
+export function as<T>(
   type: T & { new (): Serializable },
 ): FromJSONStrategy {
-  return function _fromJSONAs(
+  return function _as(
     value: JSONValue,
   ): Serializable | Serializable[] {
     if (Array.isArray(value)) {