+++ from_json/as_test.ts
@@ -1,25 +1,25 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
 import { assert, assertEquals, test } from "../test_deps.ts";
-import { fromJSONAs } from "./as.ts";
+import { as } from "./as.ts";
 import { Serializable } from "../serializable.ts";
 import { SerializeProperty } from "../serialize_property.ts";
 
 test({
-  name: "fromJSONAs revives using `fromJSON` as type",
+  name: "as revives using `fromJSON` as type",
   fn() {
     class Test extends Serializable {
       @SerializeProperty()
       test = true;
     }
     const testObj = new Test();
-    assertEquals(fromJSONAs(Test)({ test: true }).test, testObj.test);
-    assert(fromJSONAs(Test)({ test: true }) instanceof Test);
+    assertEquals(as(Test)({ test: true }).test, testObj.test);
+    assert(as(Test)({ test: true }) instanceof Test);
   },
 });
 
 test({
-  name: "fromJSONAs works in nested properties",
+  name: "as works in nested properties",
   fn() {
     class Test1 extends Serializable {
       @SerializeProperty("test_one")
@@ -28,7 +28,7 @@ test({
 
     class Test2 extends Serializable {
       @SerializeProperty(
-        { serializedKey: "test_two", fromJSONStrategy: fromJSONAs(Test1) },
+        { serializedKey: "test_two", fromJSONStrategy: as(Test1) },
       )
       test2 = new Test1();
     }
@@ -49,13 +49,13 @@ test({
 });
 
 test({
-  name: "fromJSONAs works with arrays of objects",
+  name: "as works with arrays of objects",
   fn() {
     class Test extends Serializable {
       @SerializeProperty("a_property")
       test = true;
     }
-    const array: Test[] = fromJSONAs(Test)(
+    const array: Test[] = as(Test)(
       [{ a_property: "v1" }, { a_property: "v2" }],
     );
     assertEquals(array.length, 2);